+++ docs/CHANGELOG.md
@@ -7,15 +7,23 @@
 ## Catalog, Lambdas
 !-->
 
-# 3.4.0dev - YYYY-MM-DD
+# unreleased - YYYY-MM-DD
 ## Python API
-* [Added] Support `successors` property in workflows config ([#1939](https://github.com/quiltdata/quilt/pull/1939))
 
 ## CLI
 
 ## Catalog, Lambdas
+* [Added] Generate and resolve Quilt package URIs ([#1935](https://github.com/quiltdata/quilt/pull/1935))
+
+# 3.3.0rc2 - 2020-12-04
+## Python API
+* [Added] Support `successors` property in workflows config ([#1939](https://github.com/quiltdata/quilt/pull/1939))
+
+## Catalog, Lambdas
 * [Added] Fill package metadata by dropping a JSON file into metadata controls ([#1950](https://github.com/quiltdata/quilt/pull/1950))
 * [Changed] `top_hash`-based package routes (timestamp routes are still supported in the same way) ([#1938](https://github.com/quiltdata/quilt/pull/1938))
+* [Added] Copy packages from one bucket to another ([#1932](https://github.com/quiltdata/quilt/pull/1932))
+* [Fixed] Invalid workflow when workflow is required, and there is no default workflow ([#1936](https://github.com/quiltdata/quilt/pull/1936))
 
 # 3.3.0rc1 - 2020-11-25
 ## Python API